+++ src/pharmpy/model.py
@@ -13,7 +13,6 @@ Definitions
 """
 
 import copy
-import re
 from pathlib import Path
 
 import sympy
@@ -69,6 +68,7 @@ class Model:
         model.name = self.name
         model.dependent_variable = self.dependent_variable
         model.estimation_steps = self.estimation_steps
+        model.datainfo = self.datainfo
         try:
             model.database = self.database
         except AttributeError:
@@ -184,24 +184,6 @@ class Model:
 
     def read_raw_dataset(self, parse_columns=tuple()):
         raise NotImplementedError()
-
-    def bump_model_number(self, path='.'):
-        """If the model name ends in a number increase it to next available file
-        else do nothing.
-        """
-        path = Path(path)
-        name = self.name
-        m = re.search(r'(.*?)(\d+)$', name)
-        if m:
-            stem = m.group(1)
-            n = int(m.group(2))
-            while True:
-                n += 1
-                new_name = f'{stem}{n}'
-                new_path = (path / new_name).with_suffix(self.filename_extension)
-                if not new_path.exists():
-                    break
-            self.name = new_name
 
     def create_symbol(self, stem, force_numbering=False):
         """Create a new unique variable symbol